# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ZYVIO global social, Battle and AI platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /storage/uploads/request-url:
    post:
      operationId: requestMediaUploadUrl
      tags: [storage]
      summary: Request a presigned image or video upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MediaUploadRequest" }
      responses:
        "200":
          description: Presigned upload URL
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MediaUploadResponse" }
  /storage/uploads/complete:
    post:
      operationId: completeMediaUpload
      tags: [storage]
      summary: Confirm a direct-to-storage media upload
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MediaUploadCompleteRequest" }
      responses:
        "200":
          description: Uploaded media
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MediaAttachment" }
        "409":
          description: Uploaded object is missing or its metadata does not match the declared upload
  /storage/uploads/{attachmentId}:
    delete:
      operationId: cancelMediaUpload
      tags: [storage]
      summary: Cancel and delete an unattached owned media upload
      parameters:
        - name: attachmentId
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Upload and its object were deleted
        "404":
          description: Upload not found, not owned, or already attached
        "409":
          description: Upload could not be cancelled
...
components:
...
  schemas:
...
    MediaAttachment:
      type: object
      required: [id, url, mediaType, contentType, size, originalName]
      properties:
        id: { type: string }
        url: { type: string }
        mediaType: { type: string, enum: [image, video] }
        contentType: { type: string }
        size: { type: integer }
        originalName: { type: string }
    MediaUploadRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name: { type: string, minLength: 1, maxLength: 255 }
        size: { type: integer, minimum: 1 }
        contentType: { type: string }
    MediaUploadResponse:
      type: object
      required: [uploadURL, objectPath, attachmentId, metadata]
      properties:
        uploadURL: { type: string, format: uri }
        objectPath: { type: string }
        attachmentId: { type: string }
        metadata:
          type: object
          required: [name, size, contentType, mediaType]
          properties:
            name: { type: string }
            size: { type: integer }
            contentType: { type: string }
            mediaType: { type: string, enum: [image, video] }
    MediaUploadCompleteRequest:
      type: object
      required: [attachmentId]
      properties:
        attachmentId: { type: string }
...