# lib/api-zod/src/generated/types/mediaUploadResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ZYVIO global social, Battle and AI platform API
 * OpenAPI spec version: 0.1.0
 */
import type { MediaUploadResponseMetadata } from './mediaUploadResponseMetadata';

export interface MediaUploadResponse {
  uploadURL: string;
  objectPath: string;
  attachmentId: string;
  metadata: MediaUploadResponseMetadata;
}